Feature/product

+++ src/main/java/store/buzzbook/core/repository/product/BookAuthorRepository.java
@@ -3,5 +3,5 @@ package store.buzzbook.core.repository.product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import store.buzzbook.core.entity.product.BookAuthor;
 
-public interface BookAuthorRepository extends JpaRepository<BookAuthor, Long> {
+public interface BookAuthorRepository extends JpaRepository<BookAuthor, Integer> {
 }
